requirements.txt
# # install torch and torchvision first:
# # conda install -y pytorch torchvision pytorch-cuda=12.1 -c pytorch -c nvidia
typedparser
packg
visiontext

auto-gptq  # auto-gptq for llama2 4bit model
einops
iopath
omegaconf
opencv-python-headless
pycocoevalcap
pycocotools
timm
webdataset
open-clip-torch
attrs
numpy
joblib
importlib_resources
loguru
natsort
pyyaml
python-dotenv
pathspec
zstandard
tqdm
requests
lxml
ipython
matplotlib
jupyter
notebook
ipdb
nltk
pandas
pyinstrument
nvidia-ml-py3
psutil
h5py
chardet
spacy
lmdb
msgpack
deepdiff
torchmetrics
transformers
sentence-transformers
einops-exts
cached_path
dill
base58
more-itertools
pytest
